Refactor io and structure module into structure submodules
