+++ src/protein_quest/structure/__init__.py
@@ -1,0 +1,1 @@
+"""Package dealing with gemmi.Structure objects and related functionality."""
