+++ src/protein_quest/structure/uniprot.py
@@ -1,0 +1,197 @@
+"""UniProt extraction and injection helpers for structures."""
+
+import logging
+from collections import defaultdict
+from collections.abc import Generator
+
+import gemmi
+
+from protein_quest.structure.types import Pdb2UniprotMapping, StructRefSeq
+
+logger = logging.getLogger(__name__)
+
+
+def struct_ref_seqs_columns_to_records(struct_ref_seqs_columns: dict[str, list[str | int]]) -> Generator[StructRefSeq]:
+    """Convert `_struct_ref_seq` columns into collapsed alignment records.
+
+    Args:
+        struct_ref_seqs_columns: Struct ref seqs columns.
+            As returned by `gemmi.Structure.make_mmcif_block().get_mmcif_category("_struct_ref_seq.")`.
+
+    Yields:
+        StructRefSeq records for each unique UniProt accession and chain combination."""
+    struct_ref_seqs_rows = [
+        dict(zip(struct_ref_seqs_columns.keys(), vals, strict=False))
+        for vals in zip(*struct_ref_seqs_columns.values(), strict=False)
+    ]
+
+    grouped_rows: dict[tuple[str, str], list[dict[str, str]]] = defaultdict(list)
+    for row in struct_ref_seqs_rows:
+        grouped_rows[(row["pdbx_db_accession"], row["pdbx_strand_id"])].append(row)
+
+    for (accession, chain_id), rows in grouped_rows.items():
+        starts = [int(row["db_align_beg"]) for row in rows]
+        ends = [int(row["db_align_end"]) for row in rows]
+        uniprot_start = min(starts)
+        uniprot_end = max(ends)
+        aligned_residue_count = sum(end - start + 1 for start, end in zip(starts, ends, strict=False))
+        reference_span = uniprot_end - uniprot_start + 1
+        sequence_identity = aligned_residue_count / reference_span
+        yield StructRefSeq(
+            uniprot_accession=accession,
+            uniprot_start=uniprot_start,
+            uniprot_end=uniprot_end,
+            chain_id=chain_id,
+            sequence_identity=sequence_identity,
+            aligned_residue_count=aligned_residue_count,
+        )
+
+
+def _matching_struct_ref_seqs(structure: gemmi.Structure, accessions: set[str]) -> list[StructRefSeq]:
+    block = structure.make_mmcif_block(gemmi.MmcifOutputGroups(False, struct_ref=True))
+    struct_ref_seqs_columns = block.get_mmcif_category("_struct_ref_seq.")
+    struct_ref_seqs = struct_ref_seqs_columns_to_records(struct_ref_seqs_columns)
+    return [struct_ref_seq for struct_ref_seq in struct_ref_seqs if struct_ref_seq.uniprot_accession in accessions]
+
+
+def _group_struct_ref_seqs_by_chain(struct_ref_seqs: list[StructRefSeq]) -> dict[str, list[StructRefSeq]]:
+    grouped_struct_ref_seqs: dict[str, list[StructRefSeq]] = defaultdict(list)
+    for struct_ref_seq in struct_ref_seqs:
+        grouped_struct_ref_seqs[struct_ref_seq.chain_id].append(struct_ref_seq)
+    return grouped_struct_ref_seqs
+
+
+def _select_best_struct_ref_seq(struct_ref_seqs: list[StructRefSeq]) -> StructRefSeq:
+    # Fast path: one UniProt for this chain needs no sorting.
+    if len(struct_ref_seqs) == 1:
+        return struct_ref_seqs[0]
+    # Deterministic ranking: max aligned residues, then alphabetical accession.
+    return min(struct_ref_seqs, key=lambda s: (-s.aligned_residue_count, s.uniprot_accession))
+
+
+def structure_to_uniprot(structure: gemmi.Structure) -> Pdb2UniprotMapping:
+    """Extract chain-to-UniProt mapping from a structure."""
+    chain_uniprots: set[tuple[str, str]] = set()
+    for entity in structure.entities:
+        uniprots = entity.sifts_unp_acc
+        subchains = entity.subchains
+        for subchain in subchains:
+            for uniprot in uniprots:
+                chain_uniprots.add((subchain, uniprot))
+
+    block = structure.make_mmcif_block(gemmi.MmcifOutputGroups(False, struct_ref=True))
+    struct_ref = block.get_mmcif_category("_struct_ref.")
+    db_name_unp_idx = [i for i, db_name in enumerate(struct_ref["db_name"]) if db_name == "UNP"]
+    uniprot_accessions = [struct_ref["pdbx_db_accession"][i] for i in db_name_unp_idx]
+    struct_ref_seq = block.get_mmcif_category("_struct_ref_seq.")
+    for i, uniprot_accession in enumerate(struct_ref_seq["pdbx_db_accession"]):
+        if uniprot_accession in uniprot_accessions:
+            chain_uniprots.add((struct_ref_seq["pdbx_strand_id"][i], uniprot_accession))
+
+    return {structure.name: chain_uniprots}
+
+
+def structure2uniprot_accessions(structure: gemmi.Structure) -> set[str]:
+    """Extract UniProt accessions from a gemmi Structure object.
+
+    Logs a warning and returns an empty set if no accessions are found in structure.
+
+    Args:
+        structure: The gemmi Structure object to extract UniProt accessions from.
+
+    Returns:
+        A set of UniProt accessions found in the structure."""
+    block = structure.make_mmcif_block(gemmi.MmcifOutputGroups(False, struct_ref=True))
+    struct_ref = block.get_mmcif_category("_struct_ref.")
+    uniprot_accessions: set[str] = set()
+    if not struct_ref:
+        logger.warning("No struct_ref data category found in structure %s", structure.name)
+        return uniprot_accessions
+    for i, db_name in enumerate(struct_ref["db_name"]):
+        if db_name != "UNP":
+            continue
+        pdbx_db_accession = struct_ref["pdbx_db_accession"][i]
+        uniprot_accessions.add(pdbx_db_accession)
+    if not uniprot_accessions:
+        logger.warning("No UniProt accessions found in structure %s", structure.name)
+    return uniprot_accessions
+
+
+def _append_uniprot_to_structure(
+    structure: gemmi.Structure, chain_uniprot_pairs: set[tuple[str, str]]
+) -> gemmi.Structure:
+    if not chain_uniprot_pairs:
+        return structure
+    block = structure.make_mmcif_block()
+    struct_ref = block.get_mmcif_category("_struct_ref.")
+    struct_ref_seq = block.get_mmcif_category("_struct_ref_seq.")
+    chain2entity_id: dict[str, str] = {
+        chain: entity.name for entity in structure.entities for chain in entity.subchains
+    }
+    fillable_struct_ref_seq_cols = {"align_id", "ref_id", "pdbx_strand_id", "pdbx_db_accession"}
+    for chain, uniprot_accession in chain_uniprot_pairs:
+        entity_id = chain2entity_id[chain]
+        new_id = str(len(struct_ref["id"]) + 1)
+        struct_ref["id"].append(new_id)
+        struct_ref["entity_id"].append(entity_id)
+        struct_ref["db_name"].append("UNP")
+        struct_ref["db_code"].append(None)
+        struct_ref["pdbx_db_accession"].append(uniprot_accession)
+        struct_ref["pdbx_db_isoform"].append(None)
+
+        new_seq_id = len(struct_ref_seq["align_id"]) + 1
+        struct_ref_seq["align_id"].append(new_seq_id)
+        struct_ref_seq["ref_id"].append(new_seq_id)
+        struct_ref_seq["pdbx_strand_id"].append(chain)
+        struct_ref_seq["pdbx_db_accession"].append(uniprot_accession)
+        for col in struct_ref_seq:
+            if col not in fillable_struct_ref_seq_cols:
+                struct_ref_seq[col].append(None)
+
+    block.set_mmcif_category("_struct_ref.", struct_ref)
+    block.set_mmcif_category("_struct_ref_seq.", struct_ref_seq)
+    return gemmi.make_structure_from_block(block)
+
+
+def add_uniprot_accessions2structure(
+    structure: gemmi.Structure, pdb2uniprot: Pdb2UniprotMapping | None
+) -> gemmi.Structure:
+    """Add UniProt accessions to a structure if they are missing, based on the provided pdb2uniprot mapping.
+
+    If structure has uniprot accesion that is not in `pdb2uniprot`, it will be left unchanged.
+
+    Args:
+        structure: The gemmi Structure object to add UniProt accessions to.
+        pdb2uniprot: Dictionary mapping PDB ID to set of tuples containing chain and UniProt accession.
+            If provided, will be used to inject UniProt accessions into the structure if they are missing.
+            If None, the structure is returned unchanged.
+
+    Returns:
+        A gemmi Structure object with UniProt accessions added if they were missing
+        or the structure unchanged if all accessions were already present."""
+    if not pdb2uniprot:
+        return structure
+    pdb_id = structure.name
+    if pdb_id not in pdb2uniprot:
+        logger.warning(
+            "PDB ID %s not found in pdb2uniprot mapping. Leaving structure unverified and unchanged.", pdb_id
+        )
+        return structure
+
+    known = structure_to_uniprot(structure)
+    missing = pdb2uniprot[pdb_id] - known[pdb_id]
+    if not missing:
+        return structure
+
+    if known[pdb_id]:
+        logger.warning(
+            "Structure %s has some UniProt accessions that do not match the provided mapping. "
+            "Existing: %s, Expected: %s, Missing: %s. Injecting missing accessions.",
+            pdb_id,
+            known[pdb_id],
+            pdb2uniprot[pdb_id],
+            missing,
+        )
+    else:
+        logger.info("Injecting UniProt accessions into structure %s: %s", structure.name, missing)
+    return _append_uniprot_to_structure(structure, missing)
