+++ tests/alphafold/test_confidence.py
@@ -11,8 +11,8 @@ from protein_quest.alphafold.confidence import (
     filter_out_low_confidence_residues,
     find_high_confidence_residues,
 )
-from protein_quest.io import read_structure
-from protein_quest.structure import nr_residues_in_chain
+from protein_quest.structure.chains import nr_residues_in_chain
+from protein_quest.structure.formats import read_structure
 
 
 @pytest.fixture
