--- tests/test_structure.py
@@ -1,544 +1,0 @@
-import logging
-from pathlib import Path
-from textwrap import dedent
-
-import gemmi
-import pytest
-from platformdirs import user_cache_dir
-
-from protein_quest.io import read_structure
-from protein_quest.pdbe.fetch import sync_fetch
-from protein_quest.structure import (
-    ChainNotFoundError,
-    StructRefSeq,
-    StructureMetadata,
-    nr_of_residues_in_total,
-    nr_residues_in_chain,
-    struct_ref_seqs_columns_to_records,
-    structure2uniprot_accessions,
-    structure_metadata,
-    write_single_chain_structure_file,
-)
-
-logger = logging.getLogger(__name__)
-
-
-def test_write_single_chain_structure_file_happypath(sample2_cif: Path, tmp_path: Path):
-    output_file = write_single_chain_structure_file(
-        input_file=sample2_cif,
-        chain2keep="A",
-        output_dir=tmp_path,
-        out_chain="Z",
-    )
-
-    assert output_file is not None
-    assert output_file.name == "2Y29_A2Z.cif.gz"
-    assert output_file.exists()
-    structure = read_structure(output_file)
-    assert len(structure) == 1  # One model
-    model = structure[0]
-    assert len(model) == 1  # One chain
-    chain = model[0]
-    assert chain.name == "Z"
-    assert len(chain) == 6  # 6 residues in chain Z
-
-
-def test_write_single_chain_structure_file_with_secondary_structure(sample_cif: Path, tmp_path: Path):
-    output_file = write_single_chain_structure_file(
-        input_file=sample_cif,
-        chain2keep="A",
-        output_dir=tmp_path,
-    )
-    structure = read_structure(output_file)
-    assert len(structure.helices) == 4
-    assert len(structure.sheets) == 1
-
-
-def test_write_single_chain_structure_file_already_exists(
-    sample2_cif: Path, tmp_path: Path, caplog: pytest.LogCaptureFixture
-):
-    fake_output_file = tmp_path / "2Y29_A2Z.cif.gz"
-    fake_output_file.write_text("fake content")
-    caplog.set_level(logging.INFO)
-
-    output_file = write_single_chain_structure_file(
-        input_file=sample2_cif,
-        chain2keep="A",
-        output_dir=tmp_path,
-        out_chain="Z",
-    )
-
-    assert output_file == fake_output_file
-    assert "Skipping" in caplog.text
-
-
-def test_write_single_chain_structure_file_unknown_chain(sample2_cif: Path, tmp_path: Path):
-    with pytest.raises(ChainNotFoundError):
-        write_single_chain_structure_file(
-            sample2_cif,
-            chain2keep="B",
-            output_dir=tmp_path,
-            out_chain="Z",
-        )
-
-
-def test_write_single_chain_structure_file_unknown_format(tmp_path: Path):
-    with pytest.raises(RuntimeError, match="Unknown format"):
-        write_single_chain_structure_file(
-            tmp_path / "nonexistent_file.xyz",
-            chain2keep="B",
-            output_dir=tmp_path,
-            out_chain="Z",
-        )
-
-
-@pytest.fixture
-def download_cache_dir() -> Path:
-    cache_dir = Path(user_cache_dir("protein-quest-tests"))
-    cache_dir.mkdir(parents=True, exist_ok=True)
-    return cache_dir
-
-
-def fetch_pdb_file_cached(pdb_id: str, download_cache_dir: Path) -> Path:
-    cache_fn = download_cache_dir / f"{pdb_id.lower()}.cif.gz"
-    if cache_fn.exists():
-        logger.info(f"[cache hit] Using cached file: {cache_fn}")
-        return cache_fn
-
-    logger.info(f"[cache miss] Downloading file for {pdb_id} to {cache_fn}")
-    fetched_files = sync_fetch([pdb_id], download_cache_dir, max_parallel_downloads=1)
-    fetched_file = fetched_files[pdb_id]
-    assert cache_fn == fetched_file
-    return fetched_file
-
-
-@pytest.fixture
-def lowercase_chain_cif(download_cache_dir: Path) -> Path:
-    """Big model (ribosome complex) with >36 chains"""
-    return fetch_pdb_file_cached("5KCS", download_cache_dir)
-
-
-def test_write_single_chain_structure_file_lowercase_chain(lowercase_chain_cif: Path, tmp_path: Path):
-    output_file = write_single_chain_structure_file(
-        input_file=lowercase_chain_cif,
-        chain2keep="l",
-        output_dir=tmp_path,
-        out_chain="Z",
-    )
-
-    assert output_file is not None
-    structure = read_structure(output_file)
-    assert len(structure) == 1  # One model
-    model = structure[0]
-    assert len(model) == 1  # One chain
-    chain = model[0]
-    assert chain.name == "Z"
-    assert len(chain) == 123
-
-
-@pytest.fixture
-def multi_model_cif(download_cache_dir: Path) -> Path:
-    """A multi model structure"""
-    return fetch_pdb_file_cached("2lmc", download_cache_dir)
-
-
-def test_write_single_chain_structure_file_multi_model_cif(multi_model_cif: Path, tmp_path: Path):
-    output_file = write_single_chain_structure_file(
-        input_file=multi_model_cif,
-        chain2keep="A",
-        output_dir=tmp_path,
-        out_chain="Z",
-    )
-
-    assert output_file is not None
-    structure = read_structure(output_file)
-    assert len(structure) == 1  # One model
-    model = structure[0]
-    assert len(model) == 1  # One chain
-    chain = model[0]
-    assert chain.name == "Z"
-    assert len(chain) == 59
-
-
-def test_nr_residues_in_chain(sample2_cif: Path):
-    residue_count = nr_residues_in_chain(sample2_cif, chain="A")
-
-    assert residue_count == 8
-
-
-def test_nr_residues_in_chain_wrongchain(sample2_cif: Path, caplog):
-    residue_count = nr_residues_in_chain(sample2_cif, chain="Z")
-
-    assert residue_count == 0
-    assert "Chain Z not found in" in caplog.text
-
-
-def test_structure2uniprot_accessions_present(sample2_cif: Path):
-    structure = read_structure(sample2_cif)
-    accessions = structure2uniprot_accessions(structure)
-
-    assert accessions == {"P05067"}
-
-
-def test_structure2uniprot_accessions_multiple(multi_accession_cif: Path):
-    structure = read_structure(multi_accession_cif)
-    accessions = structure2uniprot_accessions(structure)
-
-    assert accessions == {"Q13469", "P01100", "P05412"}
-
-
-def test_structure2uniprot_accessions_missing(sample_cif: Path, caplog):
-    # Empty struct_ref category to simulate missing UniProt accessions
-    structure_with_unp = read_structure(sample_cif)
-    block_without_struct_ref = structure_with_unp.make_mmcif_block(
-        gemmi.MmcifOutputGroups(True, chem_comp=False, struct_ref=False)
-    )
-    structure = gemmi.make_structure_from_block(block_without_struct_ref)
-
-    accessions = structure2uniprot_accessions(structure)
-
-    assert accessions == set()
-    assert "No UniProt accessions found in structure 3JRSB2A" in caplog.text
-
-
-def test_nr_of_residues_in_total(sample2_cif: Path):
-    structure = read_structure(sample2_cif)
-    total_residues = nr_of_residues_in_total(structure)
-
-    assert total_residues == 8
-
-
-class TestStructureMetadata:
-    @pytest.mark.parametrize(
-        "cif_fixture, expected",
-        [
-            pytest.param(
-                "sample_cif",
-                StructureMetadata(
-                    id="3JRSB2A",
-                    uniprot_accession="Q8VZS8",
-                    resolution=2.05,
-                    total_residue_count=173,
-                    is_alphafold=False,
-                    uniprot_start=8,
-                    uniprot_end=211,
-                    sequence_identity=1.0,
-                    chain_length=173,
-                    method="X-ray",
-                ),
-                id="3JRS_B2A",
-            ),
-            pytest.param(
-                "sample2_cif",
-                StructureMetadata(
-                    id="2Y29",
-                    uniprot_accession="P05067",
-                    resolution=2.3,
-                    total_residue_count=8,
-                    is_alphafold=False,
-                    uniprot_start=687,
-                    uniprot_end=692,
-                    sequence_identity=1.0,
-                    chain_length=8,
-                    method="X-ray",
-                ),
-                id="2Y29",
-            ),
-            pytest.param(
-                "af_cif",
-                StructureMetadata(
-                    id="AF-A0A0C5B5G6-F1",
-                    uniprot_accession="A0A0C5B5G6",
-                    resolution=0.0,
-                    total_residue_count=16,
-                    is_alphafold=True,
-                    uniprot_start=1,
-                    uniprot_end=16,
-                    sequence_identity=1.0,
-                    chain_length=16,
-                    method="Predicted",
-                ),
-                id="AF-A0A0C5B5G6-F1",
-            ),
-            pytest.param(
-                "nmr_cif",
-                StructureMetadata(
-                    id="1AMB",
-                    uniprot_accession="P05067",
-                    resolution=0.0,
-                    total_residue_count=28,
-                    is_alphafold=False,
-                    uniprot_start=672,
-                    uniprot_end=699,
-                    sequence_identity=1.0,
-                    chain_length=28,
-                    method="NMR",
-                ),
-                id="1AMB",
-            ),
-            pytest.param(
-                "em_cif",
-                StructureMetadata(
-                    id="8W77",
-                    uniprot_accession="P0ABE7",
-                    resolution=3.61,
-                    total_residue_count=260,
-                    is_alphafold=False,
-                    uniprot_start=23,
-                    uniprot_end=127,
-                    sequence_identity=1.0,
-                    chain_length=260,
-                    method="EM",
-                ),
-                id="8W77",
-            ),
-            pytest.param(
-                "sample_multispan_cif",
-                StructureMetadata(
-                    id="6O5I",
-                    uniprot_accession="O00255",
-                    resolution=1.240,
-                    total_residue_count=1346,
-                    is_alphafold=False,
-                    uniprot_start=1,
-                    uniprot_end=593,
-                    sequence_identity=0.816,
-                    chain_length=1346,
-                    method="X-ray",
-                ),
-                id="6O5I_multispan",
-            ),
-            pytest.param(
-                "multi_accession_cif",
-                StructureMetadata(
-                    id="1A02",
-                    uniprot_accession=None,
-                    resolution=2.7,
-                    total_residue_count=513,
-                    is_alphafold=False,
-                    uniprot_start=0,
-                    uniprot_end=0,
-                    sequence_identity=0.0,
-                    chain_length=513,
-                    method="X-ray",
-                ),
-                id="1A02_multi_accessions separate_chains",
-            ),
-            pytest.param(
-                "multi_accession_chain_cif",
-                StructureMetadata(
-                    id="1UN5",
-                    uniprot_accession="P03950",
-                    resolution=2.600,
-                    total_residue_count=131,
-                    is_alphafold=False,
-                    uniprot_start=25,
-                    uniprot_end=147,
-                    sequence_identity=119 / 123,
-                    chain_length=131,
-                    method="X-ray",
-                ),
-                id="1UN5_multi_accession_chain",
-            ),
-        ],
-    )
-    def test_cif_fixtures(self, cif_fixture: str, expected: StructureMetadata, request: pytest.FixtureRequest):
-        path = request.getfixturevalue(cif_fixture)
-        result = StructureMetadata.from_path(path)
-
-        assert result == expected
-
-    def test_multiple_accessions_warns(self, multi_accession_cif: Path, caplog: pytest.LogCaptureFixture):
-        StructureMetadata.from_path(multi_accession_cif)
-
-        message = caplog.text
-        assert "Multiple UniProt accessions found in structure" in message
-        assert "Source path:" in message
-        assert "Q13469" in message
-        assert "P01100" in message
-        assert "P05412" in message
-
-    def test_without_uniprot(self):
-        structure = gemmi.Structure()
-
-        result = structure_metadata(structure)
-
-        expected = StructureMetadata(
-            id="",
-            uniprot_accession=None,
-            resolution=0.0,
-            total_residue_count=0,
-            is_alphafold=False,
-            uniprot_start=0,
-            uniprot_end=0,
-            sequence_identity=0.0,
-            chain_length=0,
-            method="Other",
-        )
-        assert result == expected
-
-    @pytest.fixture
-    def fake_alphafill_structure(self) -> gemmi.Structure:
-        block = dedent("""\
-            data_AF-P38634-F1
-            #
-            _entry.id   AF-P38634-F1
-                    loop_
-            _software.classification
-            _software.date
-            _software.description
-            _software.name
-            _software.pdbx_ordinal
-            _software.type
-            _software.version
-            other              ?                    'Structure prediction' AlphaFold 1 package v2.0
-            other              ?                    'Secondary structure'  dssp      2 library 4
-            'model annotation' 2023-11-22T07:49:00Z ?                      alphafill 3 ?       2.1.0
-            #
-            """)
-        doc = gemmi.cif.read_string(block)
-        return gemmi.make_structure_from_block(doc.sole_block())
-
-    def test_alphafill(self, fake_alphafill_structure: gemmi.Structure):
-        result = structure_metadata(fake_alphafill_structure)
-
-        expected = StructureMetadata(
-            id="AF-P38634-F1",
-            uniprot_accession=None,
-            resolution=0.0,
-            total_residue_count=0,
-            is_alphafold=False,
-            uniprot_start=0,
-            uniprot_end=0,
-            sequence_identity=0.0,
-            chain_length=0,
-            method="Predicted",
-        )
-        assert result == expected
-
-
-@pytest.mark.parametrize(
-    ("struct_ref_seqs_columns", "expected_records"),
-    [
-        pytest.param(
-            {
-                "align_id": [1],
-                "ref_id": [1],
-                "pdbx_PDB_id_code": ["1ABC"],
-                "pdbx_strand_id": ["A"],
-                "seq_align_beg": [10],
-                "pdbx_seq_align_beg_ins_code": ["?"],
-                "seq_align_end": [15],
-                "pdbx_seq_align_end_ins_code": ["?"],
-                "pdbx_db_accession": ["P12345"],
-                "db_align_beg": [10],
-                "pdbx_db_align_beg_ins_code": ["?"],
-                "db_align_end": [15],
-                "pdbx_db_align_end_ins_code": ["?"],
-                "pdbx_auth_seq_align_beg": [10],
-                "pdbx_auth_seq_align_end": [15],
-            },
-            [StructRefSeq("P12345", 10, 15, "A", 1.0, 6)],
-            id="single-span",
-        ),
-        pytest.param(
-            {
-                "align_id": [1, 2],
-                "ref_id": [1, 2],
-                "pdbx_PDB_id_code": ["1ABC", "1ABC"],
-                "pdbx_strand_id": ["A", "A"],
-                "seq_align_beg": [10, 20],
-                "pdbx_seq_align_beg_ins_code": ["?", "?"],
-                "seq_align_end": [15, 24],
-                "pdbx_seq_align_end_ins_code": ["?", "?"],
-                "pdbx_db_accession": ["P12345", "P12345"],
-                "db_align_beg": [10, 20],
-                "pdbx_db_align_beg_ins_code": ["?", "?"],
-                "db_align_end": [15, 24],
-                "pdbx_db_align_end_ins_code": ["?", "?"],
-                "pdbx_auth_seq_align_beg": [10, 20],
-                "pdbx_auth_seq_align_end": [15, 24],
-            },
-            [StructRefSeq("P12345", 10, 24, "A", 11 / 15, 11)],
-            id="multi-span-single-chain",
-        ),
-        pytest.param(
-            {
-                "align_id": [1, 2],
-                "ref_id": [1, 2],
-                "pdbx_PDB_id_code": ["1ABC", "1ABC"],
-                "pdbx_strand_id": ["A", "B"],
-                "seq_align_beg": [10, 30],
-                "pdbx_seq_align_beg_ins_code": ["?", "?"],
-                "seq_align_end": [15, 35],
-                "pdbx_seq_align_end_ins_code": ["?", "?"],
-                "pdbx_db_accession": ["P12345", "P12345"],
-                "db_align_beg": [10, 30],
-                "pdbx_db_align_beg_ins_code": ["?", "?"],
-                "db_align_end": [15, 35],
-                "pdbx_db_align_end_ins_code": ["?", "?"],
-                "pdbx_auth_seq_align_beg": [10, 30],
-                "pdbx_auth_seq_align_end": [15, 35],
-            },
-            [
-                StructRefSeq("P12345", 10, 15, "A", 1.0, 6),
-                StructRefSeq("P12345", 30, 35, "B", 1.0, 6),
-            ],
-            id="multi-chain",
-        ),
-        pytest.param(
-            {
-                "align_id": [1, 2, 3],
-                "ref_id": [1, 1, 2],
-                "pdbx_PDB_id_code": ["1ABC", "1ABC", "1ABC"],
-                "pdbx_strand_id": ["A", "A", "A"],
-                "seq_align_beg": [10, 20, 100],
-                "pdbx_seq_align_beg_ins_code": ["?", "?", "?"],
-                "seq_align_end": [12, 22, 105],
-                "pdbx_seq_align_end_ins_code": ["?", "?", "?"],
-                "pdbx_db_accession": ["PAAAAA", "PAAAAA", "PBBBBB"],
-                "db_align_beg": [10, 20, 100],
-                "pdbx_db_align_beg_ins_code": ["?", "?", "?"],
-                "db_align_end": [12, 22, 105],
-                "pdbx_db_align_end_ins_code": ["?", "?", "?"],
-                "pdbx_auth_seq_align_beg": [10, 20, 100],
-                "pdbx_auth_seq_align_end": [12, 22, 105],
-            },
-            [
-                StructRefSeq("PAAAAA", 10, 22, "A", 6 / 13, 6),
-                StructRefSeq("PBBBBB", 100, 105, "A", 1.0, 6),
-            ],
-            id="equal-aligned-residues-same-chain",
-        ),
-        pytest.param(
-            {
-                "align_id": [1],
-                "ref_id": [1],
-                "pdbx_PDB_id_code": ["1ABC"],
-                "pdbx_strand_id": ["A"],
-                "seq_align_beg": [42],
-                "pdbx_seq_align_beg_ins_code": ["?"],
-                "seq_align_end": [42],
-                "pdbx_seq_align_end_ins_code": ["?"],
-                "pdbx_db_accession": ["P11111"],
-                "db_align_beg": [42],
-                "pdbx_db_align_beg_ins_code": ["?"],
-                "db_align_end": [42],
-                "pdbx_db_align_end_ins_code": ["?"],
-                "pdbx_auth_seq_align_beg": [42],
-                "pdbx_auth_seq_align_end": [42],
-            },
-            [StructRefSeq("P11111", 42, 42, "A", 1.0, 1)],
-            id="single-residue-inclusive-count",
-        ),
-    ],
-)
-def test_struct_ref_seqs_columns_to_records(
-    struct_ref_seqs_columns: dict[str, list[int | str]], expected_records: list[StructRefSeq]
-):
-    records = struct_ref_seqs_columns_to_records(struct_ref_seqs_columns)
-
-    assert sorted(records, key=lambda r: (r.uniprot_accession, r.chain_id)) == sorted(
-        expected_records,
-        key=lambda r: (r.uniprot_accession, r.chain_id),
-    )
